Refactor ScraperLib and DownloadState: streamline path handling, enhance state management, and improve logging. Remove unused DownloadState class file and update test imports for consistency.
