+++ pyproject.toml
@@ -33,3 +33,11 @@ docs = [
     "sphinx-copybutton>=0.5.2",
     "sphinx-design>=0.6.1",
 ]
+[tool.pytest.ini_options]
+minversion = "6.0"
+addopts = "-ra -q --cov=src/scraper_lib --cov-report=term-missing" # Adjust cov path if needed
+testpaths = [
+    "tests",
+]
+# This tells pytest to add 'src' to its path when running tests
+pythonpath = ["src"]
