+++ tests/test_customlogger.py
@@ -1,7 +1,4 @@
-import sys
-import os
 import logging
-sys.path.insert(0,  os.path.abspath(os.path.join(os.path.dirname(__file__), "../src")))
 from CustomLogger import CustomLogger
 
 def test_logger_creation(tmp_path):
