+++ tests/test_scraperlib.py
@@ -1,7 +1,4 @@
-import sys
-import os
 import pytest
-sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), "../src")))
 from ScraperLib import ScraperLib
